review: fix glow blend/edge coverage, dedupe vert + render-target, simplify copy
Addresses maintainer review:
- Composite uses premultiplied-over (ONE, ONE_MINUS_SRC_ALPHA) so the glow keeps its color over bright terrain instead of washing to white.
- ceil (not floor) the bloom buffer so edge players on non-multiple-of-scale maps still glow.
- Consolidate the 3 identical composite verts into shaders/shared/map-quad.vert.glsl (small-player-glow, fallout-bloom, fallout-light).
- Extract GlUtils.createRenderTarget and use it in the glow pass.
- Simplify the setting description; correct the throttle comment (toggle-off is tick-gated, deferred while paused).
