--- src/client/render/gl/shaders/fallout-bloom/composite.vert.glsl
@@ -1,11 +1,0 @@
-#version 300 es
-precision highp float;
-layout(location = 0) in vec2 aPos;
-uniform mat3 uCamera;
-uniform vec2 uMapSize;
-out vec2 vUV;
-void main() {
-  vec3 clip = uCamera * vec3(aPos, 1.0);
-  gl_Position = vec4(clip.xy, 0.0, 1.0);
-  vUV = aPos / uMapSize;
-}
