+++ src/client/render/gl/shaders/shared/map-quad.vert.glsl
@@ -1,13 +1,11 @@
 #version 300 es
 precision highp float;
-
-// Camera-projected map quad: samples the blurred aura in tile UV space so the
-// glow stays fixed to the map (no shimmer when panning/zooming).
+// Camera-projected map quad -> tile-space UV. Shared by the bloom/glow
+// composites so they sample an offscreen map-sized texture in world space.
 layout(location = 0) in vec2 aPos;
 uniform mat3 uCamera;
 uniform vec2 uMapSize;
 out vec2 vUV;
-
 void main() {
   vec3 clip = uCamera * vec3(aPos, 1.0);
   gl_Position = vec4(clip.xy, 0.0, 1.0);
