+++ src/server/Master.ts
@@ -129,7 +129,7 @@ export async function startMaster() {
     );
   });
 
-  const PORT = 3000;
+  const PORT = 4000;
   server.listen(PORT, () => {
     log.info(`Master HTTP server listening on port ${PORT}`);
   });
